refactor: rename pane_notation.rs -> pane.rs
