+++ pretty/src/lib.rs
@@ -13,7 +13,7 @@
 mod geometry;
 mod layout;
 mod notation;
-mod pane_notation;
+mod pane;
 mod pretty;
 mod style;
 
@@ -26,4 +26,4 @@ pub use self::notation::{
 
 pub use self::pretty::{Pane, PlainText, PrettyDocument, PrettyWindow};
 pub use self::style::{Color, ColorTheme, Emph, Rgb, Shade, Style};
-pub use pane_notation::{render_pane, Content, PaneError, PaneNotation, PaneSize};
+pub use pane::{render_pane, Content, PaneError, PaneNotation, PaneSize};
