+++ pretty/src/pane.rs
